+++ idea/testData/quickfix/createFromUsage/createFunction/call/abstract/explicitReceiverOfContainingClass.kt
@@ -1,0 +1,8 @@
+// "Create abstract function 'foo'" "true"
+abstract class A {
+    fun bar(b: Boolean) {}
+
+    fun test(a: A) {
+        bar(a.<caret>foo(1, "2"))
+    }
+}